Address API counctil review
API council feedback:

"UsecaseGroup.Builder has a setter for the viewport making it optional but in the UsecaseGroup constructor requires a non null view port. If this is a mandatory argument that should be non null, then it should be passed to the builder ctor."

The change:

Marked viewport nullable. Developer should be able to create a viewport without setting a viewport. Viewport is for synchronizing crop rect and UseCaseGroup group is for synchronizing use cases in general. One should be able to create a group for other purposes.

Relnote: fixed the bug where viewport is marked as non-null on UseCaseGroup. Developer should be able to create a viewport without setting a viewport.

Bug: 159174942
Test: manual test and ./gradlew bOS
Change-Id: Ie3d2ecb4559916680bfeee64a480cc52b1c6d627