Move nav_default animation out of animation sets
All of the nav_default animations are simple fades and do not need to be
inside of a animation set. Moving them out also ensures that if they are
placed in an additional set (i.e. like the new fragment state manager
does to listen for when the animation finishes) they will not flash.

Bug: 167430145
Test: Tested in sample app
Change-Id: I071e922a52301863ba2461ff05b60bb64566430c