Fix issue with showing BiometricPrompt back-to-back
Fixes an issue on some devices where using BiometricPrompt for
authentication again soon after it is dismissed causes the new prompt to
be dismissed as well. Because this solution involves adding artificial
latency to the second prompt invocation, it is currently limited to
known affected devices and Android OS versions.

Test: BiometricPrompt sample app from b/157783075 on API 29-30.
Test: Biometric integration test app on API 27-30.
Test: ./gradlew biometric:biometric:test
Test: ./gradlew biometric:biometric:connectedAndroidTest
Test: ./gradlew biometric:integration-tests:testapp:connectedAndroidTest

Bug: 157783075
Change-Id: I830402327b759f5297721e992c4f3e617fc94d61