Backdrop polishing
Some minor changes in BackdropScaffold:
* Updated some docs to use "scaffold" instead of "backdrop".
* Removed the sample without a snackbar and renamed the other
sample to BackdropScaffoldSample, which is now used in the demos.
* Renamed BackdropConstants to BackdropScaffoldConstants.
* Renamed the param backdropScaffoldState to scaffoldState.
* Renamed rememberBackdropState to rememberBackdropScaffoldState and
exposed the animation clock as a parameter, defaulting to the ambient.

Test: Ran BackdropScaffoldTest
Relnote: "Renamed rememberBackdropState to rememberBackdropScaffoldState
and added an additional param for the animation clock. Renamed the param
backdropScaffoldState of BackdropScaffold to just scaffoldState. Renamed
BackdropConstants to BackdropScaffoldConstants."
Change-Id: Ib644d5364cf8890ef0c637cc2ab1e0479db1eacb