Each UseCase can hold multiple UseCaseConfig which are merged
Previously there was an initial config that got merged with the Config
from the UseCaseConfig retrieved from the CameraX singleton.

Now a UseCase will have 3 Configs:
1) An app Config which contains the Optionss from the user. Which can be
updated dynamically via calls such a setTargetRotation.
2) A Camera Config which contains the default Options that come from a
specific camera implementation. This is set when a CameraInternal is
attached to the UseCase.
3) An extended Config which contains the Options which are applied on
top of the app's Configs. These are mean for camera-extensions since
they may need to override some app settings in order to guarantee the
extension can be applied. (Currently not being set)

Additional remove the usage of CameraX.getDefaultUseCaseConfig() since
the default UseCaseConfig will be applied when the CameraInternal is
attached. This also allows for almost all of the CameraX defined
UseCases to be created without initializing a CameraX instance. Only
Preview still depends on global state.

Relnote: "<UseCase>.getTargetRotation() will return Surface.ROTATION_0
if called before being attached to a Camera instance unless a
targetRotation has been set on the Builder or UseCase."

Bug: b/167693457
Test: camera:camera-{core,camera2,extensions,lifecycle}:test
      camera:camera-{core,camera2,extensions,lifecycle}:connectedCheck

Change-Id: I80fcd8b33ce27e31fd3fc38b4f9b6e0a62b06bc5