Fix visual glitches when dismissing watch face config fragments
Previously when swiping to dismiss a config fragment you'd see a 1 frame
glitch where after the top fragment has been removed to reveal the lower
one we see the top one again briefly. This appears to be due to the
asynchronous behavior of FragmentManager#popBackStack(), using
FragmentManager#popBackStackImmediate() instead causes it to go away.

Test: Observed example watch face in emulator.
Change-Id: I59100592eb29cdd902fd9df6f4c983d7f0eb7a1c