Fix FileProvider so that multiple authorities can be used.
ProviderInfo.authority is a semicolon-separated list (See https://developer.android.com/guide/topics/manifest/provider-element#auth) but FileProvider isn't splitting on semicolons and is treating the entire string as the authority. Attempting to resolve the Uri when there are multiple authorities defined in the AndroidManifest fails because resolveContentProvider expects a single authority (eg. 'example1' or 'example2'), but a joined string (eg. 'example1;example2') is passed instead.

This bug has also been reported externally:
https://stackoverflow.com/questions/43175014/possible-to-use-multiple-authorities-with-fileprovider

Test: Patched this into a sample app to verify. Passing in either the left-hand or right-hand of the authorities (when split by ';') works. Added a unit test to create and to read using the second authority.
Change-Id: Ia404bd49d7a6cbc5676804a6f206addcf11cf995