Added CameraConfig for camera specific configs
The CameraConfig for providing camera specific configurations such as if
certain camera ids need to be filtered out or UseCases attached to the
camera need additional configs.

Added ExtensionsConfig which is an implementation of CameraConfig which
provides the configurations necessary for enabling an extension.

Extension.set/getExtension() will now sets the extenions mode and the
camera filter, but the CameraUseCaseAdapter doesn't modify the UseCases
or the used CameraInternal.

Bug: b/154939118
Test: ./gradlew camera:camera-extensions:connectedCheck

Change-Id: I2d90d120f94a1981095a4bf6345dd044e4083805