Removed return value from onPointerEvent(...)
This is the 2nd CL in a 2 CL change to make
consumption of pointer event data mutable.
The first was CL 1441975.

RelNote: "Breaking change: removed the return value from PointerInputFilter.onPointerEvent(...) given that the only value that should be able to be changed in pointer events is consumption data.
Instead of returning data from PointerInputFilter.onPointerEvent(...), now you can just mutate the consumption data of the PointerEvents passed in."

Bug: 168815763
Test: ./gradlew compose:ui:ui:test
Test: ./gradlew compose:ui:ui:connectedCheck

Change-Id: I6acd06e56ab49c8ca932ff7c2d35a517a412e2d2