Identity Credential: Always use software implementation for testing.
In general we want to run tests against the software implementation
since hardware-based Identity Credential implementations are tested
against CTS and VTS and the bulk of the code in the Jetpack is the
software implementation. This also helps avoid whatever bugs or
flakiness that may exist in hardware implementations.

Also make it easy for Identity Credential Jetpack developers to
switch to using hardware-backed implementations. This is necessary
when making changes to the Jetpack API and the code responsible for
forwarding the calls to the Android platform hardware-backed APIs.

Bug: 164480361
Test: All Identity Credential Jetpack tests pass (SW and HW impls)
Relnote: N/A
Change-Id: I284c33777ffe6c6cd1a6fda2fb495f62faffb330