Moves Icon to the Material library
The tinting and the size is Material specific, so it should not be in Foundation.

Bug: b/158572605
Test: ./gradlew updateApi
Relnote: "androidx.compose.foundation.Icon has been moved to androidx.compose.material.Icon. You can also use the Image component / Modifier.paint() with a Painter if you do not want to use the Material library."
Change-Id: I9f6222326427cebacde10562cde99b9ebff2490f