Move ui.test to compose.ui.test.
I have created stubs that are redirecting to the new package which
should give me time to migrate all the tests. There is still ui/ui-test
module which purpose will be to exist for just a single release to point
to the new module via the API stubs that will navigate the developers
towards the migration.

I am also considering ui-test-rule module to isolate Junit stuff. This
should happen before I do the major refactor.

All of this is temporary and will be removed once I migrated rest of the
tests (and maybe split the ui-test-rule).

Bug: 171218089
Test: Tests pass locally
Relnote: "androidx.ui.test moved to androidx.compose.ui.test"
Change-Id: I9ffdb165d49e8d136b58cc4e32599a4a1d5b169e