LaunchedEffect is the new LaunchedTask
Rename LaunchedTask to LaunchedEffect to match the new SideEffect and
DisposableSideEffect APIs. No-args LaunchedEffect is not permitted to
encourage explicit consideration of coroutine restart conditions.

Update usages of LaunchedTask to LaunchedEffect. Change produceState
parameter names to be consistent with the, "subject" naming convention.

Relnote: "LaunchedTask was renamed to LaunchedEffect for consistency
with the SideEffect and DisposableEffect APIs. LaunchedEffect with no
subject params is not permitted in order to encourage best practices."

Test: SideEffectTests.kt

Change-Id: Ifd3d4f3b529b3956915c99096eef3fb3108b2b61