CameraController: only mirroring takePicture() when the metadata is not set explicitly.
API review feedback:

"So if the user doesn't want mirroring, they get it no matter what.  This should respect the user's choices, I think - if you want to set mirroring when no metadata was provided, that's more reasonable, but that should still be documented."

This CL changes the behavior to only mirroring the output when the field is not set by the app.

- In ImageCapture.Metadata, add an internal API for whether the mirroring has been set by the app.
- In CameraController, only mirror when it's not set by the app.

Bug: 161377155
Test: manual test and ./gradlew bOS
Change-Id: I7d4e7941f7f25203ad4107d5bbe37cfeeaa68fba