Add highlightedComplicationId to RenderParameters
This optional member allows you to request a screenshot with an
individual complication highlighted.

We also introduce androidx.wear.complications.data.IdAndComplication
and rename the wireformat version for clarity. In addition more of
the core watch face code uses the new style ComplicationData.

Test: Presubmit & the example watch face works as expected

Relnote: Add an optional highlightedComplicationId parameter to RenderParameters which allows you to request highlighting of a single complication in screenshots.
Change-Id: I66ce937929c863627c45291901b945945a21d89f