Enable the CameraX test app to swap Camera impls
This change modifies the CameraXActivity and OpenGLActivity so that a
test harness (or adb) can launch and configure an alternate camera
implementation at launch time. Since CameraX is configured per-process,
this configuration step will throw an exception if the specified
configuration is different from the currently configured implementation.
This should avoid accidental failures where an explicit launch option
could otherwise be ignored.

Test: Built and ran core test app locally
Test: Built and ran core test app using adb
Test: Build and ran core test app using alternate extras
Test: Attempted to run core test app after configuring with extras.
Change-Id: I49ec71ff5cdcf9c93afab77e41ea753718cdd22a