Change watchface init at request of WCS team
We are now able to pre-register the details of the interactive
watchface to be constructed when wallpaper service creates the
engine, reducing the period during which black frames may be
displayed.

Test: Presubmit
Relnote: Added InteractiveWatchFaceWcsClient and WatchFaceControlClient.getOrCreateWallpaperServiceBackedInteractiveWatchFaceWcsClient which either gets an existing instance or creates it once the wallaper service has connected and made the engine.
Change-Id: Id666e8c785c8cd803ff445dda4e45f40db65cee9