Set animation's monotonic state in PreviewAnimationClock
We need to set animations' monotonic state to false to be able to change the
time back and forth in animation preview without unsubscribing the animation
when it ends. That's done in TransitionAnimationClockObserver#parse(), but
this method now needs to be used by other tools that don't want to modify
the monotonic state. Therefore, we now are setting the monotonic state of
animations in PreviewAnimationClock#subscribe().

Bug: N/A
Test: PreviewAnimationClockTest should pass
Change-Id: I5ab29f3738bf4ac0e061afb20ff2abcad5aed31a