Update withTransaction test with changes in RoomDatabase#beginTransaction()
In aosp/1477951 RoomDatabase#beginTransaction() was updated to conditionally call beginTransactionNonExclusive() but the delegate in withTransaction_beginTransaction_error was not updated to also throw when that method is invoked causing this test to be 'flaky' do the failures in some API versions but not in other. This CL fixes just that by also throwing in beginTransactionNonExclusive().

Bug: 172206093
Test: SuspendingQueryTest in API > 16
Relnote: N/A
Change-Id: I6ddb012864b8af6c9d704b21a912690cad080c10