Preserve PagingState across generations if no hint received
This change resolves an issue where getRefreshKey may not be called if
the most recent generation was invalidated before it had a chance to
load data and receive a hint. This makes it impossible to preserve
anchorPosition in cases where invalidates may happen quickly.

Relnote: "Fixed an issue where getRefreshKey does not get called on
  subsequent generations of PagingSource when invalidations occur rapidly."
Fixes: 170027530
Test: ./gradlew paging:paging-common:test
Change-Id: I4546070c31ed73fd285754707b8b480cba206b93