Set clip data when sharing streams using ShareCompat
The mechanism in Intent.createChooser() to migrate clip data and flags
from the target intent only works if the target intent actually contains
the necessary flags and clip data. If that's not the case the content
preview in the share sheet won't work because it won't be able to access
the content.
With this change the share intent contains the necessary clip data items
and flags by the time Intent.createChooser() is called.

Relnote: "Set clip data and grant uri read permission when sharing streams using ShareCompat"
Bug: 173137936
Test: ShareCompatTest
Change-Id: I4aa318b146206e4e5cd028b7e3798c9d895d8324