Reduce CameraControllerFragmentTest flaking
(1) Remove the state change on the FragmentScenario
The FragmentScenario.launchInContainer automatically start the fragment.
Extra call on the moveToState might stop/start the fragment,
it  will unbind the use cases, and then bind again.

(2) Call CameraX.shutdown() after the test to ensure the camera is closed before it starts the next test.

Test: manually run CameraControllerFragmentTest
Change-Id: I87e40afba69e15d6faf950b29125a53d013940f1