getOrCreateWallpaperServiceBackedInteractiveWatchFaceWcsClient user style fix
Currently for the client there is a chicken and egg problem for
the userStyle parameter, you can't construct a UserStyle without
the schema.  We change this by allowing the user to pass in a
Map<String, String>.

We also add a toMap() method on UserStyle and a constructor
accepting Map<String, String> to help with custom seralization
of UserStyles.

Test: Presubmit
Relnote: Changed WatchFaceControlClient.getOrCreateWallpaperServiceBackedInteractiveWatchFaceWcsClient to accept a Map<String, String> instead of UserStyle because it's hard to create a UserStyle without knowing the schma which you can only get after the client has been created.
Change-Id: Iea02ada7595d6e0272a94644c470330ed17378c1