Allow remote to send NotLoading.Incomplete via LoadStateUpdate
Due to remote loads being completely async with local insert, it's no
longer possible to guarantee that local insert will finish after remote
load. This leads to situations where it's possible for remote load state
to get stuck in "Loading" state forever.

This change enables RemoteMediator to fire off LoadState events to
update to NotLoading.Incomplete, but comes with the caveat that
CombinedLoadState.mediator.refresh is now no longer
sufficient to check for displaying loadstate in UI as it may set
NotLoading before PagingSource.load finishes (and presents to the UI).

In previous versions this was worked around by resetting the Mediator
loadstate on new generations, but this is no longer accurate since those
calls don't necessarily get canceled.

Fixes: 173717820
Test: ./gradlew paging:paging-common:test
Change-Id: I2f1236fcb37777d059de3a53bbcd2a7aa038d479