Follow selection handle in TextFieldScroll
TextFieldScroll followed cursor changes for collapsed
cursor. This CL updates TextFieldScroll to follow
selection  (non-collapsed) as well.

The horizontal scrolling does not work as intended since
selection has to be updated to understand touch down
happening when touch is outside of TextField bounds.

Vertical scrolling works close to intended, however there
are more details to be handled on the selection side.

Test: N/A

Bug: 174195490
Change-Id: I6b82c86a57cd68437dc8aa88430f14510a1b1e1b