Don't call Resolver.asMemberOf for statics
KSPResolver throws an exception if asMemberOf is called with an invalid
parameter. This CL updates the wrapper guard against statics.

I'm not adding a catch all as room shouldn't really be doing this except
for statics (e.g. we wouldn't call asMemberOf if we already knew it is
not enclosed in the given type).

While writing the test, i noticed that XTypeElement implementation with
KSP does not return static fields. We actually don't need them in room
but java implementation does return them hence i've added them to KSP
as well. We might want to change that in XTypeElement in the future as
it is unnecsesary work.

Bug: 160322705
Test: KSAsMemberOfTest
Change-Id: I315eeb5346931f2936acf0df01f0f39e5de252cf