Enhance variance inheritance for overrides
This CL fixes another case on variance inheritance for overrides where
if there is an overridden method w/ star projected type argument,
we should not use start projection but instead pick covariant (out).

To reproduce it, i've expanded method spec helper test to cover methods
that are not overridden. Unfortunately, it revelaled an issue where
if a kotlin interface overrides another interface with generics, it is
not possible to generate kotlin override for it if the type argument is
a final type. (b/174313780). Only one test case hits it so I've added
a way to ignore them for now (we cannot fix it until room generates
kotlin sources).

I've also hit some KSP bugs:
https://github.com/google/ksp/issues/175
Overrides might fail when overriding with generics. Luckily,
findOverridee method finds it so I've added a workaround by calling
findOverridee. We could possibly only call it instead of calling both
but resolver bug will be eventually fixed in KSP hence I kept that call.

https://github.com/google/ksp/issues/174
Since we call findOverridee for properties now, it might trigger another
issue in KSP. Added a try catch for it now.

https://github.com/google/ksp/issues/173
KSP might throw when resolving deserialized type alias descriptors. Added
a try catch for it where we hit the issue. (reading annotations from an
element that also has @Throws annotation).

https://github.com/google/ksp/issues/164
This issue also affects get getJvmName for which we have safeGetJvmName.
In this CL, it happens when finding overridee for getter/setter methods in
java. Added a try catch for it for now.

Bug: 174313780
Bug: 160322705
Test: MethodSpecHelperTest
Change-Id: Ie6347f46027c317ed7784e8d2a0c834012fb4b52