Require ComplicationsUserStyleSetting to modify complications
This is because the new API for WCS has a pull rather than a push
model for this data and we need to ensure WCS is up to date.

The only exception is the complication bounds which you can still
change, and the excpection is WCS till pull this data each time
it's needed.

Relnote: With the exception of bounds you now have to use ComplicationsUserStyleSetting to modify complications, this is to ensure the OS is kept in synch.
Test: Presubmit
Change-Id: I8dc5db6f349e7ad4995f9ea7a3e12fe31b560c7e