Update SimpleJavaVersion to handle dot splits with non digit characters.
Certain JDK versions are simplified to either only include mayor and patch or minor and patch within the same dot split, e.g. '15+13' or '14.1-ea' which would result in the SimpleJavaVersion not correctly identifying the Java version as valid and thus causing Room to not be an incremental annotation processor. This changes updates the parser to handle those cases.

Bug:171387388
Test: SimpleJavaVersionTest
Relnotes=Fix an issue where Room would not correctly identify the JDK version to enable incremental APT.

Change-Id: I7a249a29eb901c49c883de8dcecb81c7cbf5116c