Export script now maintains only the 'external' framework folder.
All files maintained by this script are moving to that folder in the
framework codebase.

This makes it much easier to maintain the export script. Prior to this
CL, the export script had an allowlist of framework files to leave
alone. This list was badly out of date and hard to maintain in
lockstep with framework.

Java package paths and the overall API is unaffected.

Bug: 162450968
Test: Ran export
Change-Id: Idad82220a69a39e4ab8267d2c04581311976c567