Fix the compat library for display modes.
We've learned a lot about Android TV architecture in this area over
the last year and the original concept is flawed. HDMI offers no way
to determine what the 'native' mode is. In addition, the system
properties are dynamic based on the current display mode. In addition,
an app developer mostly needs getMode, which was not originally
provided.

This fixes a lot of flaws about the assumptions we made early on that
went into the initial implementation.

RelNote: Deprecated ModeCompat.isNative which is replaced by DisplayCompat.getMode.

Test: ./gradlew core:core:test
Change-Id: Iefa7170574f98c26114b143e7c01e25bda3880e8