Fixed handshake issue in `CarAppService`
The newly-introduced `Session` caused some NPEs because the onHandshakeComplete
callback was setting some data into a Session/CarContext instance that has yet
to be created until onAppCreate is called.

This change makes AppInfo to be lazily-initialized and also fixed a minor bug
in the min-api retrieval logic, by default to the current API if the min-api metadata
key is not found.

Relnote: n/a
Test: ./gradlew car:app:app:CarAppServiceTest & ./gradlew car:app:app:AppInfoTest
Bug: 171424518

Change-Id: I28db055beee93b7135439ad2888290a53ff31b7c