Remove XDeclaredType
Java treats declared types differently while for kotlin, everything is a
KSType. As XProcessing was modeled after the java APIs we use, we
initially created XDeclared type and it only had 1 method to get type
arguments.
This became a problem in the KSP implementation where it is not always
trivial to decide whether a type should be declared or not.

This CL removes XDeclaredType into XType, which in practice means adding
typeArguments property to XType.
Internal java implementation still keeps JavacDeclaredType, which simply
extends JavacType to provide type arguments. Other java types simply
return empty list for type arguments. Keeping JavacDeclaredType still
made sense as internally, it is handy to have the typeMirror with type
declared type even though Room does not care about it.

Similarly on the KSP side, I've renamed KspDeclaredType to
DefaultKspType (similar to DefaultJavacType). Even thought that
abstraction does not have any important implementation, it helps not
having default implementations in KspType for more tricky APIs (e.g. we
need to ensure subclasses of KspType implement copyWithNullability APIs,
which means KspType should stay abstract).

With this CL, `asDeclaredType` and `isDeclaredType` methods are also
removed. There were few places they are used and I've updated that logic
to match to the same behavior (in TypeAdapterStore).

Fixes: 175417726
Test: existing tests

Change-Id: I193eb2b7f7c70cb5d530411b802c22986a808fc2