Move rc01 libraries that have been cut to alpha01
Activity, Fragment, Lifecycle, SavedState, and
Transition have all had their release branches
snapped for their rc01 releases.

Update the versions of all of those libraries to
alpha01 of their next version to allow for
development to start.

Moved AppCompat and WorkManager to depend on
prebuilts to avoid a dependency on alpha builds
of other libraries.

Test: ./gradlew bOS
Change-Id: I08a91eb3168d85d485570fc5d93e702cc90b88c3