Add nullable versions of ColumnTypeAdapters for KSP
With javac, Room's type converters support nullable types (e.g.
java.lang.Boolean) using its boxed primtives.

This does not work well with KSP where we have proper type nullability
which is considered in type equality.

This CL adds nullable versions of boxed ColumnTypeAdapters.
It is limited to KSP and also still generates the same code for declared
types even when we know the type is not nullable. This decision can be
changed in the future to generate better code (we'll need to do that for
kotlin codegen).

I've limited these null aware adapters to KSP because in javac, type
comparison ignores nullability.

I've also fixed a possible bug in String and byte[] converters. Right
now, when reading values, they always call cursor getString/Blob which
happens to return null but appearantly it is not guarateed so now we
generate code that checks for null.

Bug: 160322705
Bug: 176104406
Test: BasicColumnTypeAdaptersTests are now run with KSP as well and also
test for nullable boxed types
Test: SQLTypeAffinityTest

Change-Id: Ib4e74f54fdb5ab83e58a8287c5324f434cb0fddd